+++ src/Assets/Bugsnag/Editor/BugsnagEditor.cs
@@ -6,11 +6,20 @@ using System.IO;
 using BugsnagUnity;
 using UnityEditor.Callbacks;
 using System.Linq;
+using System;
 
 namespace BugsnagUnity.Editor
 {
     public class BugsnagEditor : EditorWindow
     {
+        // The kotlin Version here needs to match the one in the rake build file. Both should reflect what the android notifier is using.
+        private const string ANDROID_DEPS_XML = "<dependencies><androidPackages><repositories><repository>https://mvnrepository.com/artifact/org.jetbrains.kotlin/kotlin-stdlib</repository></repositories><androidPackage spec=\"org.jetbrains.kotlin:kotlin-stdlib:1.4.32\"></androidPackage></androidPackages></dependencies>";
+
+        private const string EDM_MENU_ITEM = "Window/Bugsnag/Enable EDM4U Support";
+
+        private static string EDMDepsFilePath = "/Bugsnag/Editor/BugsnagAndroidDependencies.xml";
+
+        private static string KotlinLibsDirPath = "/Bugsnag/Plugins/Android/Kotlin";
 
         private bool _showBasicConfig = true;
 
@@ -21,13 +30,33 @@ namespace BugsnagUnity.Editor
         private Vector2 _scrollPos;
 
 
+        [MenuItem(EDM_MENU_ITEM,false,1)]
+        private static void ToggleEDM()
+        {
+            if (IsEDMEnabled())
+            {
+                DisableEDM();
+            }
+            else
+            {
+                EnableEDM();
+            }
+        }
+
+        [MenuItem(EDM_MENU_ITEM, true)]
+        private static bool ToggleEDMValidate()
+        {
+            Menu.SetChecked(EDM_MENU_ITEM, IsEDMEnabled());
+            return true;
+        }
+
         private void OnEnable()
         {
             titleContent.text = "Bugsnag";
             CheckForSettingsCreation();
         }
 
-        [MenuItem("Window/Bugsnag/Configuration")]
+        [MenuItem("Window/Bugsnag/Configuration",false,0)]
         public static void ShowWindow()
         {
             CheckForSettingsCreation();
@@ -224,5 +253,135 @@ namespace BugsnagUnity.Editor
             File.WriteAllLines(pbxPath, lines.ToArray());
         }
 #endif
+
+        public static void EnableEDM()
+        {
+            try
+            {
+                EditDepsFile(true);
+                UpdateKotlinLibraryImportSettings(false);
+                AssetDatabase.Refresh();
+            }
+            catch (Exception e)
+            {
+                Debug.LogError("Error enabling Bugsnag EDM4U support: " + e.Message);
+            }
+
+            if (IsEDMEnabled())
+            {
+                ReportEDMSuccess("Bugsnag EDM4U support successfully enabled.\n\nPlease restart Unity before building.");
+            }
+            else
+            {
+                ReportEDMError("Error enabling Bugsnag EDM4U support.\n\nPlease check the console for error messages");
+            }
+        }
+
+        public static void DisableEDM()
+        {
+            try
+            {
+                EditDepsFile(false);
+                UpdateKotlinLibraryImportSettings(true);
+                AssetDatabase.Refresh();
+            }
+            catch (Exception e)
+            {
+                Debug.LogError("Error disabling Bugsnag EDM4U support: " + e.Message);
+            }
+
+            if (!IsEDMEnabled())
+            {
+                ReportEDMSuccess("Bugsnag EDM4U support successfully disabled.\n\nPlease restart Unity before building.");
+            }
+            else
+            {
+                ReportEDMError("Error disabling Bugsnag EDM4U support.\n\nPlease check the console for error messages");
+            }
+        }
+
+        private static void ReportEDMSuccess(string msg)
+        {
+            BugsnagEDMPopup.Open(msg);
+            Debug.Log(msg);
+        }
+
+        private static void ReportEDMError(string msg)
+        {
+            BugsnagEDMPopup.Open(msg);
+            Debug.LogError(msg);
+        }
+
+        private static void UpdateKotlinLibraryImportSettings(bool active)
+        {
+            foreach (var lib in GetKotlinLibs())
+            {
+                lib.SetCompatibleWithPlatform(BuildTarget.Android, active);
+                lib.SaveAndReimport();
+            }
+        }
+
+        private static void EditDepsFile(bool create)
+        {
+            var path = Application.dataPath + EDMDepsFilePath;
+            if (create)
+            {
+                File.WriteAllText(path, ANDROID_DEPS_XML);
+            }
+            else
+            {
+                File.Delete(path);
+                File.Delete(path + ".meta");
+            }
+        }
+
+        private static List<PluginImporter> GetKotlinLibs()
+        {
+            var kotlinLibs = new List<PluginImporter>();
+            foreach (var libPath in Directory.GetFiles(Application.dataPath + KotlinLibsDirPath, "*.jar"))
+            {
+                kotlinLibs.Add((PluginImporter)AssetImporter.GetAtPath(libPath.Replace(Application.dataPath, "Assets")));
+            }
+            return kotlinLibs;
+        }
+
+        private static bool IsEDMEnabled()
+        {
+            if (!File.Exists(Application.dataPath + EDMDepsFilePath))
+            {
+                return false;
+            }
+            foreach (var lib in GetKotlinLibs())
+            {
+                if (lib.GetCompatibleWithPlatform(BuildTarget.Android))
+                {
+                    return false;
+                }
+            }
+            return true;
+        }
+    }
+
+    public class BugsnagEDMPopup : EditorWindow
+    {
+        private static string _msg;
+
+        public static void Open(string msg)
+        {
+            _msg = msg;
+            BugsnagEDMPopup window = ScriptableObject.CreateInstance<BugsnagEDMPopup>();
+            window.position = new Rect(Screen.width / 2, Screen.height / 2, 200, 110);
+            window.ShowPopup();
+        }
+
+        void OnGUI()
+        {
+            var style = EditorStyles.wordWrappedLabel;
+            style.alignment = TextAnchor.MiddleCenter;
+            GUILayout.Space(5);
+            EditorGUILayout.LabelField(_msg,style );
+            GUILayout.Space(5);
+            if (GUILayout.Button("ok")) this.Close();
+        }
     }
 }